Copy of the original lambda-launcher-dns.py
Problem with updating Route53 record upon instance launch
is that CFN loses track of what the DNS record looks like
and won't delete it when deleting the stack.

Instead we have to update the stack parameters with
the IP of the newly launched instance and let CFN update
the DNS record.